Index processing

+++ FunDBLib/MetaData/MetaField.cs
@@ -12,14 +12,14 @@ namespace FunDBLib.MetaData
 
         public PropertyInfo Property { get; set; }
 
-        public byte Length { get; set; }
+        public int Length { get; set; }
         
         public MetaField()
         {
             
         }
 
-        public MetaField(string name, EnumFieldTypes fieldType, PropertyInfo property, byte length)
+        public MetaField(string name, EnumFieldTypes fieldType, PropertyInfo property, int length)
         {
             Name = name;
             FieldType = fieldType;
